requirements.txt
# ── Core API ─────────────────────────────────────────────────────
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.9.2
pydantic-settings==2.5.2
python-dotenv==1.0.1

# ── LangChain + LangGraph ─────────────────────────────────────────
langchain==0.3.7
langchain-openai==0.2.0
langchain-community==0.3.7
langgraph==0.2.39

# ── Vector Store ──────────────────────────────────────────────────
pinecone-client==5.0.1
langchain-pinecone==0.2.0

# ── Cache ─────────────────────────────────────────────────────────
redis==5.1.1
langchain-community==0.3.7   # includes RedisSemanticCache

# ── LLM ──────────────────────────────────────────────────────────
openai==1.51.0
tiktoken==0.8.0

# ── Guardrails ────────────────────────────────────────────────────
presidio-analyzer==2.2.355
presidio-anonymizer==2.2.355

# ── Observability ─────────────────────────────────────────────────
structlog==24.4.0

# ── Testing ──────────────────────────────────────────────────────
pytest==8.3.3
pytest-asyncio==0.24.0
httpx==0.27.2
